requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.1
serpapi>=0.2.4
pytest>=8.0.0
